cc: Don't implicitly convert size_t to unsigned int.
They may have different sizes, and this is causing a compiler warning
to be issued on win_x64.

e:\b\build\slave\win\build\src\cc\resources\resource_provider.h(69) : warning C4267: 'return' : conversion from 'size_t' to 'unsigned int', possible loss of data

TBR=piman
BUG=

Review URL: https://codereview.chromium.org/22022002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@215469 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b0e5a5e89b1279fe6868cb9eb0fdf7accbf410fd

+++ cc/resources/resource_provider.h
@@ -66,7 +66,7 @@ class CC_EXPORT ResourceProvider {
   WebKit::WebGraphicsContext3D* GraphicsContext3D();
   int max_texture_size() const { return max_texture_size_; }
   GLenum best_texture_format() const { return best_texture_format_; }
-  unsigned num_resources() const { return resources_.size(); }
+  size_t num_resources() const { return resources_.size(); }
 
   // Checks whether a resource is in use by a consumer.
   bool InUseByConsumer(ResourceId id);
